// apps/backend/src/types/category.ts
export enum CategoryAccountFlow {
  In = 'I',
  Out = 'S',
}

export enum CategoryTypeMove {
  Sale = 'V',
  Expense = 'G',
}

export enum CategoryStatus {
  Active = 'A',
  Inactive = 'E',
}

export enum CategoryTypeId {
  Store = 'T',
  Bank = 'B',
  Liquidator = 'L',
  Standard = 'S',
  Detraction = 'D',
  Multiple = 'M',
}